Add new

+++ Auto_save_problem_for_CTFd.py
@@ -1,10 +1,11 @@
 # -*- coding: UTF-8 -*-
 # @Author   : ERROR404
-# @Version  : 1.0
+# @Version  : 1.2
 import requests,os,ssl,html,urllib.request,time,shutil
 
 ssl._create_default_https_context = ssl._create_unverified_context
-DEBUG = False # 将此处置位将会输出部分except的异常详细信息！
+DEBUG = True # 将此处置位将会输出部分except的异常详细信息！
+Save_File = True # 您是否想要自动化存储题目附件？
 
 Headers={
     'accept': '*/*',
@@ -13,7 +14,8 @@ Headers={
     'sec-fetch-mode': 'cors',
     'sec-fetch-site': 'same-origin',
     'user-agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_2) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.130 Safari/537.36',
-    'x-requested-with': 'XMLHttpRequest'
+    'x-requested-with': 'XMLHttpRequest',
+    'csrf-token': '6056c8f6b07709306c2a6eae4f366717c2799e5c762da696bbdffd4425428b53'
 }
 
 Cookies={
@@ -56,16 +58,23 @@ def Save(url,race_name):
     Log_content  = '本程序在 %s 启动，开始自动储存题目!\n' % (time.strftime("%Y-%m-%d %H:%M:%S", time.localtime())) 
     save_success_num = 0
 
+    if not save_file:
+        Log_content += '[!] 本次存储中，附件存储功能已关闭！这将极大地提升存储速度，您需要稍后手动存储附件！'
+        print('\033[0;31m' + '[!] 本次存储中，附件存储功能已关闭！这将极大地提升存储速度，您需要稍后手动存储附件！' + '\033[0m')
+
     race_path = '~/Desktop/CTF_question/' + race_name + '/'
 
     try:
         API_url = url + 'api/v1/challenges'
         All_challenges = requests.get(API_url,headers=Headers,cookies=Cookies)
         challenges_list = All_challenges.json()['data']
-    except:
+    except Exception as e:
         if DEBUG:
             print(e)
-        print('\033[0;31m' + '[-] 与服务器链接失败或服务器返回值异常!请检查Cookie或URL是否正确设置!' + '\033[0m')
+        if( ('message' in All_challenges.json().keys()) and (All_challenges.json()['message'] == 'You don\'t have the permission to access the requested resource. It is either read-protected or not readable by the server.') ):
+            print('\033[0;31m' + '[-] 已与服务器建立链接!但比赛仍未开始!' + '\033[0m')
+        else:
+            print('\033[0;31m' + '[-] 与服务器链接失败或服务器返回值异常!请检查Cookie或URL是否正确设置!' + '\033[0m')
         exit(0)
 
     for challenges in challenges_list:
@@ -131,32 +140,34 @@ def Save(url,race_name):
                 print('\033[0;31m' + '[-] %s 方向新增题目 %s 失败!请尝试再次存储!' % (category,problem_name) + '\033[0m')
                 continue
 
-        if len(problem['files']) != 0 and upgrade == False:
-            remaining_file_number = len(problem['files'])
-            Log_content += '[!] 题目 %s 共计有 %d 个附件!\n' % (problem_name,remaining_file_number)
-            print('\033[0;33m' + '[!] 题目 %s 共计有 %d 个附件!' % (problem_name,remaining_file_number) + '\033[0m')
-            for index in range(remaining_file_number):
-                Log_content += '[!] 开始存储题目 %s 的第 %d 个附件!\n' % (problem_name,remaining_file_number)
-                print('\033[0;33m' + '[!] 开始存储题目 %s 的第 %d 个附件!' % (problem_name,remaining_file_number) + '\033[0m')
-                if (save_file(url,challenge_path + get_filename(problem['files'][index]),problem['files'][index])):
-                    Log_content += '[!] 题目 %s 的第 %d 个附件存储成功!\n' % (problem_name,index+1)
-                    print('\033[0;32m' + '[!] 题目 %s 的第 %d 个附件存储成功!' % (problem_name,index+1) + '\033[0m')
-                    Save_fail = False
-                else:
-                    shutil.rmtree(os.path.expanduser(challenge_path))
-                    Log_content += '[!] 题目 %s 的第 %d 个附件存储失败!题目文件夹已被清除!!\n' % (problem_name,index+1)
-                    print('\033[0;31m' + '[!] 题目 %s 的第 %d 个附件存储失败!题目文件夹已被清除!' % (problem_name,index+1) + '\033[0m')
-                    Save_fail = True
-                    break
-            if Save_fail :
-                Log_content += '[-] %s 方向新增题目 %s 失败!请尝试再次存储!\n' % (category,problem_name)
-                print('\033[0;31m' + '[-] %s 方向新增题目 %s 失败!请尝试再次存储!' % (category,problem_name) + '\033[0m')
-                continue
+        if save_file:
+            if len(problem['files']) != 0 and upgrade == False:
+                remaining_file_number = len(problem['files'])
+                Log_content += '[!] 题目 %s 共计有 %d 个附件!\n' % (problem_name,remaining_file_number)
+                print('\033[0;33m' + '[!] 题目 %s 共计有 %d 个附件!' % (problem_name,remaining_file_number) + '\033[0m')
+                for index in range(remaining_file_number):
+                    Log_content += '[!] 开始存储题目 %s 的第 %d 个附件!\n' % (problem_name,index+1)
+                    print('\033[0;33m' + '[!] 开始存储题目 %s 的第 %d 个附件!' % (problem_name,index+1) + '\033[0m')
+                    if (save_file(url,challenge_path + get_filename(problem['files'][index]),problem['files'][index])):
+                        Log_content += '[!] 题目 %s 的第 %d 个附件存储成功!\n' % (problem_name,index+1)
+                        print('\033[0;32m' + '[!] 题目 %s 的第 %d 个附件存储成功!' % (problem_name,index+1) + '\033[0m')
+                        Save_fail = False
+                    else:
+                        shutil.rmtree(os.path.expanduser(challenge_path))
+                        Log_content += '[!] 题目 %s 的第 %d 个附件存储失败!题目文件夹已被清除!!\n' % (problem_name,index+1)
+                        print('\033[0;31m' + '[!] 题目 %s 的第 %d 个附件存储失败!题目文件夹已被清除!' % (problem_name,index+1) + '\033[0m')
+                        Save_fail = True
+                        break
+                if Save_fail :
+                    Log_content += '[-] %s 方向新增题目 %s 失败!请尝试再次存储!\n' % (category,problem_name)
+                    print('\033[0;31m' + '[-] %s 方向新增题目 %s 失败!请尝试再次存储!' % (category,problem_name) + '\033[0m')
+                    continue
 
         try:
             challenge_exist_file = open(os.path.expanduser(challenge_exist_file_path),'w')
             challenge_exist_file.write("此文件由自动存题脚本生成，如需更新题目描述，请删除本文件！")
             challenge_exist_file.close()
+            save_success_num = save_success_num + 1
         except Exception as e:
             if DEBUG:
                 print(e)
@@ -183,6 +194,7 @@ if __name__ == "__main__":
     # 请从浏览器获取该字段然后填在这里哦~
     url = ''
     Cookies['session'] = ''
+    # Cookies['Other'] = 'Other'
     #以上为用户预定义部分↑
 
     if race_name != '' and Cookies['session'] != '' and url != '':
